♻️ Fold account-deletion into the user feature and the profile route
account-deletion is a sub-concern of user (data + row mutations only),
not a feature. The UI and the schedule/cancel actions are used solely by
the profile page, so they become route-private; the reaper is composed
in the house-keeping route, which may import user and billing freely —
this is what keeps the user feature free of billing imports.
