+++ apps/web/src/app/[locale]/(space)/settings/profile/components/delete-account-button.tsx
