+++ apps/web/src/app/[locale]/(space)/settings/profile/components/profile-picture.tsx
