+++ apps/web/src/features/user/account-deletion/data.ts
@@ -3,6 +3,25 @@ import "server-only";
 import { prisma } from "@rallly/database";
 import { upcomingScheduledEventWhere } from "@/features/scheduled-event/utils";
 
+export async function findUsersScheduledForRemoval({
+  cutoff,
+  excludeUserIds,
+  limit,
+}: {
+  cutoff: Date;
+  excludeUserIds: string[];
+  limit: number;
+}) {
+  return prisma.user.findMany({
+    where: {
+      deletedAt: { lt: cutoff },
+      id: { notIn: excludeUserIds },
+    },
+    select: { id: true, email: true, customerId: true },
+    take: limit,
+  });
+}
+
 // Counts are limited to columns with a userId index. Vote and comment
 // counts would scan participants/comments on an unindexed user_id, so the
 // dialog mentions those without numbers.
